requirements.txt
solrpy
pysolr